update ssl script to check if nginx is running in container

+++ docker/nginx/ssl_generation_run_from_virtualmachine.sh
@@ -2,9 +2,17 @@
 
 # run docker container with
 # docker compose up -d
-# then run this script with www.yourdomain.com
+# then run this script with yourdomain.com
 
-printf "%s" "enter domain name: "
+if [ -z "$(docker ps -q -f name=nginx)" ]; then 
+        echo "Nginx is not running, please run docker compose up -d, and try again."
+        exit
+fi 
+
+
+
+
+printf "%s" "enter domain name(ie google.com, no www): "
 read server_name
 docker compose run --rm certbot certonly --webroot \
     --webroot-path /var/www/certbot \
